refactor GameStateRenderer, split functionality by game state
